This is an implementation of depth count in stage 1: it tells you the terminating depth, specifically. This is one of the ways I'm thinking about making certain on-demand parsing stops (it can overrun if the final depth is greater than zero). It may also help us with actual streaming (knowing where the chunk is going to stop in advance).

It doesn't really help with performance for the existing code; and I'm unsure if it'll help with other code, either :) Posting here for posterity.

### Haswell gcc10 (Skylake)

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| instruments                      |   3442 | 112.4 | 113.0 |    0% | 370.8 | 403.6 |   -8% |
| twitterescaped                   |   8787 | 190.3 | 191.3 |    0% | 488.1 | 518.9 |   -6% |
| random                           |   7976 | 148.9 | 149.8 |    0% | 484.3 | 523.4 |   -8% |
| update-center                    |   8330 | 118.3 | 119.8 |   -1% | 329.5 | 363.3 |  -10% |
| semanticscholar-corpus           | 134271 | 112.4 | 115.0 |   -2% | 278.2 | 309.3 |  -11% |
| marine_ik                        |  46616 | 288.2 | 297.8 |   -3% | 899.9 | 934.0 |   -3% |
| citm_catalog                     |  26987 |  85.3 |  88.2 |   -3% | 289.8 | 318.6 |   -9% |
| twitter                          |   9867 |  93.7 |  97.7 |   -4% | 284.0 | 314.8 |  -10% |
| apache_builds                    |   1988 |  89.5 |  93.4 |   -4% | 298.5 | 331.1 |  -10% |
| canada                           |  35172 | 273.6 | 286.2 |   -4% | 953.6 | 976.0 |   -2% |
| mesh.pretty                      |  24646 | 171.0 | 180.8 |   -5% | 575.8 | 603.6 |   -4% |
| mesh                             |  11306 | 294.3 | 313.0 |   -6% | 989.2 | 1020.1 |   -3% |
| github_events                    |   1017 |  81.1 |  86.4 |   -6% | 259.2 | 289.6 |  -11% |
| gsoc-2018                        |  51997 |  72.8 |  77.6 |   -6% | 162.7 | 189.9 |  -16% |
| numbers                          |   2345 | 244.8 | 270.8 |  -10% | 797.8 | 836.1 |   -4% |