@jkeiser 2020-09-16
Closing since @lemire has likely seen this and I don't want to merge it :)